refactor: Score 패키지 이동
