+++ src/main/java/bowling/domain/BowlingGameFrameRecord.java
@@ -2,8 +2,6 @@ package bowling.domain;
 
 import java.util.List;
 
-import bowling.domain.frame.state.Score;
-
 public class BowlingGameFrameRecord {
     private final List<Score> scores;
