+++ src/main/java/bowling/domain/frame/state/Ready.java
@@ -2,6 +2,8 @@ package bowling.domain.frame.state;
 
 import java.util.List;
 
+import bowling.domain.Score;
+
 public class Ready implements State {
     @Override
     public State bowl(int falledPins) {
