+++ src/main/java/bowling/domain/frame/state/State.java
@@ -1,5 +1,7 @@
 package bowling.domain.frame.state;
 
+import bowling.domain.Score;
+
 public interface State {
     State bowl(int pins);
